修改bug

+++ database/migrations/2018_08_27_135726_create_article_praises_table.php
@@ -17,7 +17,7 @@ class CreateArticlePraisesTable extends Migration
             $table->increments('id');
             $table->integer('article_id')->index();
             $table->integer('user_id')->index();
-            $table->integer('ip')->index();
+            $table->bigInteger('ip')->index();
             $table->timestamps();
         });
     }
